Update nixpkgs and the way it is handled
The nixpkgs-channels repo is deprecated and the git repo is getting to
large to clone for every use so tar.gz archives of specific commits are
fetched from github instead of the entire git repo and history. User
interface remains the same however the structure of the version json
file has suffered a backwards incompatible change. Only users that wish
to up date to the latest version of a channel need download the entire
git repo now speeding up first time usage after a version change, this
can even be obviated by specifying to the update script which git commit
to update to.